Major Code Refactoring
