+++ src/App.tsx
@@ -3,7 +3,6 @@ import { BrowserRouter } from 'react-router-dom'
 import { Provider } from 'react-redux'
 
 import { Routes } from './routes/Routes'
-import store from './store/store'
 import LoadingSpinner from './components/LoadingSpinner'
 import store from './store/store'
