+++ src/components/RockPaperScissors/Button.tsx
@@ -1,0 +1,19 @@
+import React, { ReactElement } from 'react'
+import { StyledButton } from './styles/Button.styled'
+import { Option } from '../../pages/RockPaperScissors/util'
+
+type Props = {
+  option: Option
+  onClick: () => void
+  children: ReactElement
+}
+
+const Button = (props: Props) => {
+  return (
+    <StyledButton borderColor={props.option.color} onClick={props.onClick}>
+      {props.children}
+    </StyledButton>
+  )
+}
+
+export default Button
