+++ src/components/RockPaperScissors/styles/PickState.styled.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const PickStateContainer = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+`
